+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - 2.4.3 *Unreleased*
     - Fix font-size slider control’s colouring.
+    - Fix broken Advanced Help button.
+    - Sync UI with other PreviewApps.
 - 2.4.2 *16 April 2026*
     - Fix paragraph breaks in content copied from previews.
     - Reduce image placeholder sizes and change referenced path's colour.